FM-99: Move helper structs
